[feat] allow configurable evaluation function
The prompting workflow can become a lot more smooth with a custom
evaluation function. This prevents the user from tracking the completion
of each job, and then running evaluation code and copy/pasting the
numbers. With a custom evalution function + wandb, the user can start
the job and directly go to wandb to see the final metrics.